backend/requirements.txt
fastapi
uvicorn
praw
yfinance
vaderSentiment
pandas
pandas-ta
python-dotenv
pydantic-settings
requests
httpx
